Update modbot.py
Move the already_done append to the end of the loop so gdata isn't re-searching every submission. 
Also, apparently, when a video has 0 views, instead of retuning a 0 for entry.statistics.view_count, it doesn't set the view_count attribute at all. so, do deal with that I change the conditional expression for dealing with view counts. I can only hope that python doesn't process second parts of "and" booleans if the first part is false... or that we never run into an issue where a video is so new it has 0 views again.