Update modbot.py
As it turns out, the main for loop for processing the newest 100 submissions can  except for 504 gateway timeout time out, so I moved the entire loop behind a try/except so that hopefully on the next loop of the while statement the gateway is responding properly again. Is it obvious I'm new to working with generators?